requirements.txt
ultralytics==8.3.0
streamlit
torch
Pillow
gdown
